requirements.txt
scipy==1.3.3
SimpleIT==2.0.2
pyradiomics==3.0.1        # radiomic features
keras==2.2.4              # lung segmentation
scikit-image==0.16.2
